backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
email-validator>=2.1.0
google-genai
python-dotenv>=1.0.0
slowapi>=0.1.8
tiktoken>=0.5.2
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
aiofiles>=23.2
starlette>=0.36.3
httpx>=0.27.0
typing-extensions>=4.9.0
pytest>=7.4.4
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.1
pillow>=10.0.0